nav bar front + back 99%

+++ DilanDjar_Back/src/main/java/com/ffoui/DilanDjar/doas/OrderDao.java
@@ -1,6 +1,7 @@
 package com.ffoui.DilanDjar.doas;
 
 import com.ffoui.DilanDjar.entities.Order;
+import com.ffoui.DilanDjar.exceptions.ResourceNotFoundException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
@@ -18,8 +19,7 @@ public class OrderDao {
     private final RowMapper<Order> orderRowMapper = (rs, _) -> new Order(
             rs.getInt("id"),
             rs.getInt("user_id"),
-            rs.getString("email"),
-            rs.getTimestamp("order_date").toLocalDateTime()
+            rs.getString("email")
     );
 
     public List<Order> getAllOrders() {
@@ -28,17 +28,33 @@ public class OrderDao {
     }
 
     public List<Order> getOrdersByEmail(String email) {
-        String sql = "SELECT * FROM products WHERE email = ?";
-        return jdbcTemplate.query(sql, orderRowMapper, email);
+        String sql = "SELECT * FROM orders WHERE email = ?";
+        List<Order> orders = jdbcTemplate.query(sql, orderRowMapper, email);
+
+        if (orders.isEmpty()) {
+            throw new ResourceNotFoundException("Aucune commande trouvée pour l'Email : " + email);
+        }
+
+        return orders;
+    }
+
+    private boolean orderExistsById(Integer id) {
+        String checkSql = "SELECT COUNT(*) FROM orders WHERE id = ?";
+        Integer count = jdbcTemplate.queryForObject(checkSql, Integer.class, id);
+        return count != null && count > 0;
     }
 
     public boolean createOrder(Order order) {
-        String sql = "INSERT INTO orders (user_id, email, order_date) VALUES (?, ?, ?)";
-        int rowsAffected = jdbcTemplate.update(sql, order.getUserId(), order.getEmail(), order.getOrderDate());
+        String sql = "INSERT INTO orders (user_id, email) VALUES (?, ?)";
+        int rowsAffected = jdbcTemplate.update(sql, order.getUserId(), order.getEmail());
         return rowsAffected > 0;
     }
 
     public boolean deleteOrder(Integer id) {
+        if (!orderExistsById(id)) {
+            throw new ResourceNotFoundException("Commande avec l'ID : " + id + " n'existe pas");
+        }
+
         String sql = "DELETE FROM orders WHERE id = ?";
         int rowsAffected = jdbcTemplate.update(sql, id);
         return rowsAffected > 0;
